Fix startup bugs and improve observability
- Fix chainlist JSON parse: ChainId long, filter IDs > int.MaxValue
- Fix Kafka: apache/kafka image, full script paths, 9094 external port
- Fix TimescaleDB PKs: include time column, drop-and-recreate if not hypertable
- Fix continuous aggregate refresh windows to satisfy minimum bucket ratio
- Promote block/transaction logs from Debug to Information

+++ ETHTPS/ETHTPS.ChainRegistry/Clients/ChainlistNetwork.cs
@@ -5,7 +5,7 @@ namespace ETHTPS.ChainRegistry.Clients;
 public class ChainlistNetwork
 {
     [JsonPropertyName("chainId")]
-    public int ChainId { get; set; }
+    public long ChainId { get; set; }
 
     [JsonPropertyName("name")]
     public string Name { get; set; } = "";
